feat: melhoria no build
