+++ package.json
@@ -5,8 +5,10 @@
   "main": "index.js",
   "tag": "dev",
   "scripts": {
-    "build": "rimraf dist && lerna exec -- rollup -c ../../scripts/rollup.config.js && gulp --gulpfile ./scripts/gulpfile.cjs",
-    "dev": "lerna exec -- rollup -c ../../scripts/rollup.config.js -w",
+    "build:packages": "lerna exec -- rollup -c ../../scripts/packages-rollup.config.js",
+    "build:bunddle": "rollup -c ./scripts/bunddle-rollup.config.js",
+    "build": "rimraf dist &&  npm run build:packages && npm run build:bunddle",
+    "dev": "lerna exec -- rollup -c ../../scripts/packages-rollup.config.js -w",
     "start": "live-server --open=demo",
     "test": "jest --coverage",
     "test:dev": "jest --watch",
